refactor(ui): fix install-keys lint warnings and simplify UsageMeter
Move `DeprecatedBadge` from `constants.tsx` to `StatusChip.tsx` and
`usageLabel` into `UsageMeter` so each file exports only components
or only non-components (react-refresh/only-export-components). Add
`role="presentation"` to the actions column wrapper
(jsx-a11y/no-static-element-interactions).

Refactor `UsageMeter` to accept an `InstallKey` directly and absorb
the blocker/usage derivation that both call sites duplicated. Use
`cn()` for class merging in `UsageMeter` and `StatusChip`.
