+++ ui/apps/console/src/pages/install-keys/constants.tsx
@@ -5,14 +5,9 @@ import {
   HandRaisedIcon,
   ListBulletIcon,
 } from "@heroicons/react/24/outline";
-import StatusChip from "./StatusChip";
 
 type IconType = ComponentType<SVGProps<SVGSVGElement>>;
 
-/**
- * Single source of truth for the enrollment modes: the icon that carries each mode's identity, its
- * label, a one-liner of what it does, and the longer description used in the create/edit selector.
- */
 export const MODE_INFO: Record<
   string,
   { label: string; icon: IconType; summary: string; description: string }
@@ -48,12 +43,4 @@ export const MODE_INFO: Record<
 
 export function modeInfo(mode: string) {
   return MODE_INFO[mode] ?? MODE_INFO.automatic;
-}
-
-/**
- * Marks the namespace's auto-managed tenant-only registration key: a legacy path kept for
- * compatibility and slated for removal, so it carries a caution (deprecated) badge.
- */
-export function DeprecatedBadge() {
-  return <StatusChip label="Deprecated" tone="yellow" />;
 }
